move archive into separate file, add linting to makefile
